rtofs directory; lint exclude
