+++ utils/obsproc/CMakeLists.txt
@@ -1,1 +1,2 @@
+add_subdirectory(rtofs)
 add_subdirectory(applications)
