+++ utils/obsproc/rtofs/j
@@ -1,0 +1,33 @@
+UpdateCTestConfiguration  from :/home/edwardg/da/edwardg/global-workflow/sorc/gdas.cd/build/gdas-utils/DartConfiguration.tcl
+UpdateCTestConfiguration  from :/home/edwardg/da/edwardg/global-workflow/sorc/gdas.cd/build/gdas-utils/DartConfiguration.tcl
+Test project /home/edwardg/da/edwardg/global-workflow/sorc/gdas.cd/build/gdas-utils
+Constructing a list of tests
+Done constructing a list of tests
+Updating test list for fixtures
+Added 0 tests to meet fixture requirements
+Checking test dependency graph...
+Checking test dependency graph end
+test 1
+    Start 1: test_gdasapp_util_coding_norms
+
+1: Test command: /home/edwardg/da/edwardg/global-workflow/sorc/gdas.cd/build/bin/gdas-utils_cpplint.py "--quiet" "--recursive" "/home/edwardg/da/edwardg/global-workflow/sorc/gdas.cd/bundle/gdas-utils/test/../"
+1: Environment variables: 
+1:  OMP_NUM_THREADS=1
+1: Test timeout computed to be: 10000000
+1: /home/edwardg/da/edwardg/global-workflow/sorc/gdas.cd/bundle/gdas-utils/test/../obsproc/rtofs/RTOFSDataFile.cc:1:  bundle/gdas-utils/obsproc/rtofs/RTOFSDataFile.cc should include its header file bundle/gdas-utils/obsproc/rtofs/RTOFSDataFile.h  [build/include] [5]
+1: /home/edwardg/da/edwardg/global-workflow/sorc/gdas.cd/bundle/gdas-utils/test/../obsproc/rtofs/RTOFSOb.cc:2:  bundle/gdas-utils/obsproc/rtofs/RTOFSOb.cc should include its header file bundle/gdas-utils/obsproc/rtofs/RTOFSOb.h  [build/include] [5]
+1/1 Test #1: test_gdasapp_util_coding_norms ...***Failed    2.31 sec
+
+0% tests passed, 1 tests failed out of 1
+
+Label Time Summary:
+gdas-utils    =   2.31 sec*proc (1 test)
+script        =   2.31 sec*proc (1 test)
+
+Total Test time (real) =   2.32 sec
+
+The following tests FAILED:
+	  1 - test_gdasapp_util_coding_norms (Failed)
+Errors while running CTest
+Output from these tests are in: /home/edwardg/da/edwardg/global-workflow/sorc/gdas.cd/build/gdas-utils/Testing/Temporary/LastTest.log
+Use "--rerun-failed --output-on-failure" to re-run the failed cases verbosely.
